refactor: ArticleRequest -> ArticleCondition으로 이름 변경
