+++ src/main/java/com/sprint/monew/domain/article/dto/ArticleCondition.java
@@ -1,10 +1,10 @@
-package com.sprint.monew.domain.article.dto.request;
+package com.sprint.monew.domain.article.dto;
 
 import java.time.Instant;
 import java.util.List;
 import java.util.UUID;
 
-public record ArticleRequest(
+public record ArticleCondition(
     String keyword,
     UUID interestId,
     List<String> sourceIn,
