+++ src/main/java/com/sprint/monew/domain/article/repository/ArticleRepositoryCustom.java
@@ -1,7 +1,7 @@
 package com.sprint.monew.domain.article.repository;
 
 import com.sprint.monew.domain.article.dto.ArticleDto;
-import com.sprint.monew.domain.article.dto.request.ArticleRequest;
+import com.sprint.monew.domain.article.dto.ArticleCondition;
 import java.util.List;
 import java.util.UUID;
 import org.springframework.data.domain.Pageable;
@@ -9,9 +9,9 @@ import org.springframework.data.domain.Slice;
 
 public interface ArticleRepositoryCustom {
 
-  Slice<ArticleDto> getArticles(ArticleRequest condition, UUID userId, Pageable pageable);
+  Slice<ArticleDto> getArticles(ArticleCondition condition, UUID userId, Pageable pageable);
 
   List<String> findAllSourceUrl();
 
-  Long getArticleCount(ArticleRequest condition);
+  Long getArticleCount(ArticleCondition condition);
 }
